pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auditml"
version = "0.1.0"
description = "Comprehensive Privacy Auditing Toolkit for PyTorch Models"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name="Eeman Asghar"}]
dependencies = [
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "scikit-learn>=1.3.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.12.0",
  "pyyaml>=6.0",
  "click>=8.1.0",
  "tqdm>=4.65.0",
  "opacus>=1.4.0"
]

[project.scripts]
auditml = "auditml.cli:auditml"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
